refactor: WI-0296 employee page decomposition phase1
