+++ ROADMAP.md
@@ -566,3 +566,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0293 employee withholding receipt locale dynamic UI baseline(`/employee/withholding-receipt` ko/en locale copy wiring via useI18n + runtime log locale/krw formatting(ko-KR/en-US) + WI-0293 regression test)
 - WI-0294 locale dynamic UI gap fix baseline(`/employee/payslip-receipts` + `/admin/leave-calendar` ko/en locale copy wiring via useI18n + runtime log locale mapping(ko-KR/en-US) + WI-0294 regression test)
 - WI-0295 core decomposition phase1 baseline(`src/features/payroll/service.ts` year-end helper split into `year-end-calculation-helpers.ts`, `year-end-filing-artifact-helpers.ts`, `year-end-filing-submission-query-helpers.ts` + `src/app/admin/page.tsx` utility/type extraction to `page-helpers.ts` and `page-types.ts` + WI-0295 regression test)
+- WI-0296 employee decomposition phase1 baseline(`src/app/employee/page.tsx` local type/utility extraction to `src/app/employee/page-types.ts` + `src/app/employee/page-helpers.ts` with no behavior change + WI-0296 regression test)
