+++ src/app/employee/page.tsx
@@ -3,274 +3,57 @@
 import Link from "next/link";
 import { useEffect, useMemo, useState } from "react";
 
+import {
+  buildQuery,
+  calculateNetMinutes,
+  coerceNumber,
+  estimateLeaveRequestedDays,
+  firstDayOfMonthLocal,
+  formatDateTime,
+  formatDays,
+  isDevToolsEnabled,
+  lastDayOfMonthLocal,
+  matchesRequestSearch,
+  shiftDays,
+  sortRequestRowsByOption,
+  startOfLocalDay,
+  statusToTone,
+  todayEndLocal,
+  todayStartLocal,
+  toIso,
+  toLocalDateKey,
+  toLocalInputValue,
+  toTimestamp
+} from "@/app/employee/page-helpers";
+import type {
+  ApiLog,
+  AttendanceRecordDto,
+  IntegratedSubmitChecklistCard,
+  IntegratedSummaryCard,
+  LeaveBalanceDto,
+  LeaveCalendarDensity,
+  LeaveCalendarDayCell,
+  LeaveCalendarStatusTone,
+  LeaveRequestDto,
+  MobileRequestTimelineItem,
+  PreSubmitCheckItem,
+  RequestFailureCause,
+  RequestFeedbackRow,
+  RequestSearchRow,
+  RequestSearchScope,
+  RequestSortOption,
+  RequestStatusFilter,
+  ResubmitCandidate,
+  TimelineChannelFilter,
+  WorkScheduleDto
+} from "@/app/employee/page-types";
 import { EmployeeJourneyShortcutPanel } from "@/components/employee-self-service/EmployeeJourneyShortcutPanel";
 import { useSupabaseSession } from "@/lib/client/useSupabaseSession";
 import { useStickyStringState } from "@/lib/client/useStickyState";
 
-type ApiLog = {
-  id: number;
-  label: string;
-  status: number;
-  ok: boolean;
-  durationMs: number;
-  at: string;
-  body: unknown;
-};
-
-type AttendanceRecordDto = {
-  id: string;
-  employeeId: string;
-  checkInAt: string;
-  checkOutAt: string | null;
-  breakMinutes: number;
-  isHoliday: boolean;
-  notes: string | null;
-  state: "PENDING" | "APPROVED" | "REJECTED";
-};
-
-type LeaveRequestDto = {
-  id: string;
-  employeeId: string;
-  leaveType: "ANNUAL" | "SICK" | "UNPAID";
-  startDate: string;
-  endDate: string;
-  unit: "FULL_DAY" | "HALF_DAY" | "HOUR";
-  hours: number | null;
-  days: number;
-  reason: string | null;
-  decisionReason?: string | null;
-  state: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED";
-};
-
-type WorkScheduleDto = {
-  id: string;
-  employeeId: string;
-  startAt: string;
-  endAt: string;
-  breakMinutes: number;
-  isHoliday: boolean;
-  notes: string | null;
-};
-
-type LeaveBalanceDto = {
-  employeeId: string;
-  grantedDays: number;
-  usedDays: number;
-  remainingDays: number;
-  carryOverDays: number;
-  lastAccrualYear: number | null;
-  updatedAt: string;
-};
-
-type LeaveCalendarDensity = "none" | "low" | "mid" | "high";
-type LeaveCalendarStatusTone = "none" | "approved" | "pending" | "rejected" | "mixed";
-
-type LeaveCalendarDayCell = {
-  dateKey: string;
-  dayOfMonth: number;
-  inCurrentMonth: boolean;
-  isToday: boolean;
-  requestCount: number;
-  approvedCount: number;
-  pendingCount: number;
-  rejectedCount: number;
-  density: LeaveCalendarDensity;
-  tone: LeaveCalendarStatusTone;
-};
-
-type RequestFeedbackRow = {
-  id: string;
-  channel: "attendance" | "leave";
-  status: string;
-  at: string;
-  message: string;
-  tone: "ok" | "pending" | "fail";
-};
-
-type RequestFailureCause = {
-  id: string;
-  source: string;
-  message: string;
-  at: string;
-};
-
-type RequestStatusFilter = "all" | "PENDING" | "APPROVED" | "REJECTED" | "CANCELED";
-type TimelineChannelFilter = "all" | "attendance" | "leave";
-
-type MobileRequestTimelineItem = {
-  id: string;
-  channel: "attendance" | "leave";
-  status: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED";
-  at: string;
-  title: string;
-  detail: string;
-};
-
-type PreSubmitCheckItem = {
-  id: string;
-  pass: boolean;
-  label: string;
-  detail: string;
-};
-
-type IntegratedSummaryCard = {
-  key: string;
-  label: string;
-  value: string;
-  detail: string;
-  tone: "ok" | "pending" | "fail" | "info";
-};
-
-type ResubmitCandidate = {
-  key: string;
-  channel: "attendance" | "leave";
-  recordId: string;
-  status: "REJECTED" | "CANCELED";
-  at: string;
-  reason: string;
-  summary: string;
-};
-
-type IntegratedSubmitChecklistCard = {
-  key: string;
-  label: string;
-  passCount: number;
-  totalCount: number;
-  ready: boolean;
-  detail: string;
-  targetSectionId: string;
-};
-
-type RequestSearchScope = "all" | "request_id" | "status" | "content";
-type RequestSortOption = "pending_first" | "latest_desc" | "oldest_asc" | "status";
-
-type RequestSearchRow = {
-  key: string;
-  channel: "attendance" | "leave";
-  requestId: string;
-  status: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED";
-  at: string;
-  summary: string;
-  detail: string;
-  pendingHours: number;
-};
-
 const LEAVE_CALENDAR_WEEKDAYS = ["일", "월", "화", "수", "목", "금", "토"] as const;
 
-function isDevToolsEnabled() {
-  const raw = process.env.NEXT_PUBLIC_FLOWHR_DEV_TOOLS ?? "";
-  const normalized = raw.trim().toLowerCase();
-  return normalized === "1" || normalized === "true" || normalized === "yes" || normalized === "on";
-}
-
-function toLocalInputValue(value: Date) {
-  const adjusted = new Date(value.getTime() - value.getTimezoneOffset() * 60_000);
-  return adjusted.toISOString().slice(0, 16);
-}
-
-function todayStartLocal() {
-  const now = new Date();
-  return toLocalInputValue(new Date(now.getFullYear(), now.getMonth(), now.getDate(), 9, 0, 0));
-}
-
-function todayEndLocal() {
-  const now = new Date();
-  return toLocalInputValue(new Date(now.getFullYear(), now.getMonth(), now.getDate(), 18, 0, 0));
-}
-
-function firstDayOfMonthLocal() {
-  const now = new Date();
-  return toLocalInputValue(new Date(now.getFullYear(), now.getMonth(), 1, 0, 0, 0));
-}
-
-function lastDayOfMonthLocal() {
-  const now = new Date();
-  return toLocalInputValue(new Date(now.getFullYear(), now.getMonth() + 1, 0, 23, 59, 0));
-}
-
-function toIso(value: string) {
-  return new Date(value).toISOString();
-}
-
-function toLocalDateKey(value: Date | string) {
-  const parsed = typeof value === "string" ? new Date(value) : value;
-  const adjusted = new Date(parsed.getTime() - parsed.getTimezoneOffset() * 60_000);
-  return adjusted.toISOString().slice(0, 10);
-}
-
-function startOfLocalDay(value: Date) {
-  return new Date(value.getFullYear(), value.getMonth(), value.getDate(), 0, 0, 0, 0);
-}
-
-function shiftDays(value: Date, days: number) {
-  const shifted = new Date(value);
-  shifted.setDate(shifted.getDate() + days);
-  return shifted;
-}
-
-function coerceNumber(value: string, fallback = 0) {
-  const parsed = Number(value);
-  if (!Number.isFinite(parsed)) {
-    return fallback;
-  }
-  return parsed;
-}
-
-function buildQuery(params: Record<string, string | undefined>) {
-  const search = new URLSearchParams();
-  for (const [key, value] of Object.entries(params)) {
-    if (value === undefined || value.trim() === "") {
-      continue;
-    }
-    search.set(key, value);
-  }
-  const query = search.toString();
-  return query.length > 0 ? `?${query}` : "";
-}
-
-function formatDateTime(value: string | null) {
-  if (!value) {
-    return "-";
-  }
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return value;
-  }
-  return parsed.toLocaleString("ko-KR");
-}
-
-function formatDays(value: number) {
-  return Number.isInteger(value) ? String(value) : value.toFixed(2);
-}
-
-function toTimestamp(value: string | null) {
-  if (!value) {
-    return 0;
-  }
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return 0;
-  }
-  return parsed.getTime();
-}
-
 const ATTENDANCE_NOTE_PRESETS = ["퇴근 누락 정정", "출근 시각 정정", "휴게시간 정정"] as const;
-
-function calculateNetMinutes(input: { checkInAt: string; checkOutAt: string | null; breakMinutes: number }) {
-  const checkInMs = new Date(input.checkInAt).getTime();
-  if (Number.isNaN(checkInMs)) {
-    return null;
-  }
-  if (!input.checkOutAt) {
-    return null;
-  }
-  const checkOutMs = new Date(input.checkOutAt).getTime();
-  if (Number.isNaN(checkOutMs) || checkOutMs <= checkInMs) {
-    return null;
-  }
-  const grossMinutes = Math.round((checkOutMs - checkInMs) / 60_000);
-  return grossMinutes - Math.max(0, Math.trunc(input.breakMinutes));
-}
 
 function formatDeltaMinutes(deltaMinutes: number) {
   if (deltaMinutes === 0) {
@@ -288,100 +71,6 @@ function formatDeltaMinutes(deltaMinutes: number) {
   }
   const sign = deltaMinutes > 0 ? "+" : "-";
   return `${sign}${parts.join(" ")}`;
-}
-
-function statusToTone(status: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED") {
-  if (status === "APPROVED") {
-    return "ok";
-  }
-  if (status === "PENDING") {
-    return "pending";
-  }
-  return "fail";
-}
-
-function statusSortRank(status: "PENDING" | "APPROVED" | "REJECTED" | "CANCELED") {
-  if (status === "PENDING") {
-    return 0;
-  }
-  if (status === "REJECTED") {
-    return 1;
-  }
-  if (status === "CANCELED") {
-    return 2;
-  }
-  return 3;
-}
-
-function sortRequestRowsByOption(rows: RequestSearchRow[], sortOption: RequestSortOption) {
-  return [...rows].sort((left, right) => {
-    if (sortOption === "latest_desc") {
-      return toTimestamp(right.at) - toTimestamp(left.at);
-    }
-    if (sortOption === "oldest_asc") {
-      return toTimestamp(left.at) - toTimestamp(right.at);
-    }
-    if (sortOption === "status") {
-      const statusDiff = statusSortRank(left.status) - statusSortRank(right.status);
-      if (statusDiff !== 0) {
-        return statusDiff;
-      }
-      return toTimestamp(right.at) - toTimestamp(left.at);
-    }
-    const pendingDiff = Number(right.status === "PENDING") - Number(left.status === "PENDING");
-    if (pendingDiff !== 0) {
-      return pendingDiff;
-    }
-    if (right.status === "PENDING" && left.status === "PENDING") {
-      const waitDiff = right.pendingHours - left.pendingHours;
-      if (waitDiff !== 0) {
-        return waitDiff;
-      }
-    }
-    return toTimestamp(right.at) - toTimestamp(left.at);
-  });
-}
-
-function matchesRequestSearch(scope: RequestSearchScope, query: string, row: RequestSearchRow) {
-  if (!query) {
-    return true;
-  }
-  const normalizedRequestId = row.requestId.toLowerCase();
-  const normalizedStatus = row.status.toLowerCase();
-  const normalizedContent = `${row.summary} ${row.detail} ${row.channel}`.toLowerCase();
-
-  if (scope === "request_id") {
-    return normalizedRequestId.includes(query);
-  }
-  if (scope === "status") {
-    return normalizedStatus.includes(query);
-  }
-  if (scope === "content") {
-    return normalizedContent.includes(query);
-  }
-  return `${normalizedRequestId} ${normalizedStatus} ${normalizedContent}`.includes(query);
-}
-
-function estimateLeaveRequestedDays(input: {
-  startDate: string;
-  endDate: string;
-  unit: "FULL_DAY" | "HALF_DAY" | "HOUR";
-  hoursInput: string;
-}) {
-  if (input.unit === "HALF_DAY") {
-    return 0.5;
-  }
-  if (input.unit === "HOUR") {
-    return Math.max(0, coerceNumber(input.hoursInput)) / 8;
-  }
-
-  const startMs = new Date(input.startDate).getTime();
-  const endMs = new Date(input.endDate).getTime();
-  if (Number.isNaN(startMs) || Number.isNaN(endMs) || endMs < startMs) {
-    return 0;
-  }
-  const dayMs = 24 * 60 * 60 * 1000;
-  return Math.max(1, Math.ceil((endMs - startMs + 1) / dayMs));
 }
 
 function extractErrorMessage(body: unknown) {
